requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
scipy>=1.9.0
numpy>=1.23.0
matplotlib>=3.5.0
scikit-learn>=1.1.0
tqdm>=4.64.0
opencv-python>=4.7.0
flask>=2.3.0
flask-cors>=4.0.0